requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
coverage==7.2.5
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1.7
django-bootstrap-v5==1.0.11
django-cleanup==7.0.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-taggit==3.1.0
Faker==18.4.0
gunicorn==20.1.0
idna==3.4
jsonfield==3.1.0
oauthlib==3.2.2
parameterized==0.9.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
reretry==0.11.8
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4
sqlparse==0.4.3
swapper==1.3.0
tzdata==2023.3
urllib3==1.26.15
